pyproject.toml
[tool.poetry]
name = "quantfinnance"
version = "1.0.0"
description = "Algorithmic Trading System with Backtesting and Paper Trading"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.0.0"
numpy = "^2.2.6"
yfinance = "^0.2.0"
matplotlib = "^3.7.0"
plotly = "^5.0.0"
alpaca-trade-api = "^3.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
pandas-ta = "*"
scipy = "^1.10.0"
mplfinance = "^0.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
